requirements.txt
aiogram>=3.3.0
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.0
aiosqlite>=0.19.0
